requirements.txt
python-dotenv
langchain
langchain-community
pinecone-client
tiktoken
langchain-groq
groq
streamlit